requirements.txt
requests
redshift_connector
pandas
python-dotenv
SQLAlchemy
ipykernel
```